@@ -1,14 +1,15 @@
 module github.com/jenkins-x/gen-crd-api-reference-docs
 
-go 1.15
+go 1.21.0
 
 require (
-	github.com/go-logr/logr v1.2.4 // indirect
-	github.com/pkg/errors v0.9.1
-	github.com/pmezard/go-difflib v1.0.0 // indirect
 	github.com/russross/blackfriday/v2 v2.1.0
-	github.com/shurcooL/sanitized_anchor_name v1.0.0 // indirect
-	k8s.io/gengo v0.0.0-20230306165830-ab3349d207d4
-	k8s.io/klog v1.0.0
-	k8s.io/klog/v2 v2.90.1 // indirect
+	k8s.io/gengo/v2 v2.0.0-20240404160639-a0386bf69313
+	k8s.io/klog/v2 v2.130.1
+)
+
+require (
+	github.com/go-logr/logr v1.4.1 // indirect
+	golang.org/x/mod v0.14.0 // indirect
+	golang.org/x/tools v0.16.1 // indirect
 )
